CubeSat NovaOne/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-dateutil==2.8.2
python-multipart==0.0.6
aiofiles==23.2.1
structlog==23.2.0
redis==5.0.1
